Update Makefile and Workfows from ECR
Why these changes are being introduced:
This repository will use an ECR repository in AWS to store the
container for the task and will publish the container automatically
via GitHub Actions.

How this addresses that need:
* Update Makefile with developer commands to simplify the method for
pushing a copy of the container to the ECR repository in our dev AWS
account
* Create three caller workflows that leverage our shared container
publishing workflows in our GitHub organization.

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-647